+++ packages/headless/src/features/redirection/redirection-slice.ts
@@ -2,6 +2,9 @@ import {createReducer} from '@reduxjs/toolkit';
 import {checkForRedirection} from './redirection-actions';
 import {getRedirectionInitialState} from './redirection-state';
 
+/**
+ * @deprecated - Please use `standaloneSearchBoxSetReducer` instead.
+ */
 export const redirectionReducer = createReducer(
   getRedirectionInitialState(),
   (builder) =>