+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.56.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.56.0...@coveo/atomic@0.56.1) (2021-08-10)
+
+
+### Bug Fixes
+
+* **atomic:** increase memory for building atomic ([#1064](https://github.com/coveo/ui-kit/issues/1064)) ([f820660](https://github.com/coveo/ui-kit/commit/f8206605097a1099b2d451041b5507c19657ba48))
+
+
+
+
+
+# [0.56.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.55.0...@coveo/atomic@0.56.0) (2021-08-10)
+
+
+### Features
+
+* **atomic:** add ripple effect to clickable facet elements ([#1058](https://github.com/coveo/ui-kit/issues/1058)) ([1a2c0ee](https://github.com/coveo/ui-kit/commit/1a2c0ee071c131f17535551e77e3a8df349d65dc))
+
+
+
+
+
+# [0.55.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.54.1...@coveo/atomic@0.55.0) (2021-08-10)
+
+
+### Features
+
+* **headless:** allow relative redirection urls, communicate state using localstorage  ([#1042](https://github.com/coveo/ui-kit/issues/1042)) ([9d8d0cb](https://github.com/coveo/ui-kit/commit/9d8d0cb922f0abf57b44bbe70e36202c168bffd3))
+
+
+
+
+
+## [0.54.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.54.0...@coveo/atomic@0.54.1) (2021-08-09)
+
+
+### Bug Fixes
+
+* **atomic:** fixed new assets directory being created on each build ([#1061](https://github.com/coveo/ui-kit/issues/1061)) ([ee9ce36](https://github.com/coveo/ui-kit/commit/ee9ce36c2c8e5f9315c5c7d8903e191823f93366))
+
+
+
+
+
 # [0.54.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.53.0...@coveo/atomic@0.54.0) (2021-08-05)