+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.14.1](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.14.0...@coveo/headless@1.14.1) (2021-08-10)
+
+
+### Bug Fixes
+
+* **headless:** various fixes for query suggest "q" param updates ([#1065](https://github.com/coveo/ui-kit/issues/1065)) ([e9b43f6](https://github.com/coveo/ui-kit/commit/e9b43f6d40ff4dfa47ae5b99ba0e31b803edc3aa))
+
+
+
+
+
+# [1.14.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.13.1...@coveo/headless@1.14.0) (2021-08-10)
+
+
+### Features
+
+* **headless:** allow relative redirection urls, communicate state using localstorage  ([#1042](https://github.com/coveo/ui-kit/issues/1042)) ([9d8d0cb](https://github.com/coveo/ui-kit/commit/9d8d0cb922f0abf57b44bbe70e36202c168bffd3))
+
+
+
+
+
+## [1.13.1](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.13.0...@coveo/headless@1.13.1) (2021-08-09)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
 # [1.13.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.12.0...@coveo/headless@1.13.0) (2021-08-05)